requirements.txt
Flask==2.0.1
lyricsgenius==3.0.1
Pillow==9.0.1
python-dotenv==0.19.2
tweepy==4.6.0
Unidecode==1.3.2
